requirements.txt
opencv-python==4.9.0.80
mediapipe==0.10.9
ultralytics==8.1.0
deepface==0.0.89
anthropic==0.25.0
sqlalchemy==2.0.28
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1
PyQt6==6.6.1
plotly==5.20.0
plyer==2.1.0
pystray==0.19.5
Pillow==10.2.0
python-dotenv==1.0.1
